move PathAssignments
